pyproject.toml
[tool.poetry]
name = "umls_thesaurus"
version = "0.1.0"
description = ""
authors = ["Sean Davis <seandavi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
google-cloud-storage = "^1.28.1"
google-cloud-bigquery = "^1.24.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
sphinx = "^3.0.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
